cambio nombre de packages Verificador --> Corrector
