+++ src/main/java/edu/fiuba/algo3/modelo/Corrector/CorrectorConOrden.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo.Verificador;
+package edu.fiuba.algo3.modelo.Corrector;
 
 import edu.fiuba.algo3.modelo.Opciones.Opciones;
